requirements.txt
pydantic
pillow
click
rich